Switch id columns to SERIAL data type for autoincrement to account for SQLite default behaviour

+++ resources/postgres_migrations/002-create-releases-table.up.sql
@@ -1,10 +1,10 @@
 CREATE TABLE releases
 (
-    id          bigint primary key,
+    id          serial primary key,
     group_id    text        not null,
     artifact_id text        not null,
     version     text        not null,
     created_ts  timestamptz not null,
-    build_id    bigint,
+    build_id    integer,
     foreign key (build_id) references builds (id)
 );
